update headers util
