+++ app/routes/settings+/profile.connections.tsx
@@ -20,7 +20,7 @@ import {
 	providerNames,
 } from '#app/utils/connections.tsx'
 import { prisma } from '#app/utils/db.server.ts'
-import { pipeHeaders } from '#app/utils/remix.server.js'
+import { pipeHeaders } from '#app/utils/headers.server.js'
 import { makeTimings } from '#app/utils/timing.server.ts'
 import { createToastHeaders } from '#app/utils/toast.server.ts'
 import { type Info, type Route } from './+types/profile.connections.ts'
